gtk3: fix build
We depend on the newly fixed cairo-gobject to build. There's also a weird
interaction between my OS (Fedora 21) and Anaconda, which for some reason
installs its own version of <math.h> (??!?!). The active combination of gate
macros ends up with M_PI not getting defined, which was breaking compilation
in a few files.