Big round of updates.
I'm now compiling in a centos6 docker container because compiling on
my Fedora machine ended up injecting ABI dependencies on very recent
versions of glibc. (See

  http://www.win.tue.nl/~aeb/linux/misc/gcc-semibug.html

.) This exposed a bunch of problems with what I was doing. I
transitioned my packages to depend on the Conda version of libffi,
but it turns out that there are a couple of issues with how libffi
is currently packaged in Conda that made this a biiiig pain in the
ass. (See

  https://github.com/conda/conda-recipes/issues/272

.)