pgplot: that was harder than I thought
So, the way that FORTRAN stores strings is incompatible with Conda's
prefix-substitution approach, so my previous attempt was not helpful. I got
around this by adding a C shim (which took me way too long to get working, but
that's another story) that exports its value into FORTRAN.

I'm a little worried because Conda isn't reporting detections of a hardcoded
path in libpgplot.a, which I think should be there. But CASA stuff uses
libpgplot.so anyway.