dbus-client: fill in $PREFIX/var/lib/dbus/machine-id
I can't quite reproduce the situation but it appears that in certain cases
this file has to exist for the DBus client library to be happy. Fortunately
conda provides a hook so that we can generate a unique machine-id upon
installation of the package, if we build and include the additional
dbus-uuidgen program.

I'm a little worried about whether we'll need to also include dbus-launch,
etc; we'll cross that bridge when we come to it.