py3cairo: import
This hasn't been updated in forever and so needs a bunch of patches to build,
which I've stolen from Debian. The patch that adds the Region object doesn't
work, even though it looks like everything is being done OK. I can't figure
out what's going on there and there's only so much time I want to spend on this.