recipes/at-spi2-core: update to 2.18.3
While the configure script says it requires intltool >= 0.40.0, we build
just fine without it, so add a patch to eliminate that requirement.

We do however, require X. libexec/at-spi2-registryd requires libXtst, which I
wasn't packaging before. It may not really be necessary, but it's not a big
deal to include it.