recipes/pgplot: build for OS X
Here we venture into the scary land of compiling Fortran code on OS X. It
looks like Homebrew has a package that provides GFortran and generally Does
The Right Thing so hopefully this isn't so bad. But I worry about the binary
compatibility situation.

Also I converted this package to use modern-xorg-stack.