Work on portability of glib
First of all, we were embedding deps on the Brew /usr/local version of libintl. It
looks like I've really got to add my own gettext package, so let's go and do that.

Then stock glib recently added a requirement of OS X 10.9 or higher. That's pretty
aggressive. It's all for one GIo notification subsystem, so I've just disabled it.
We'll see if anything breaks.