recipes/modern-xorg-stack: patch xmkmf files for Conda's special setup
The support files for imake/xmkmf have some bits that hardcode the name of the
library directory to "lib64" on 64-bit systems. Normally this is sensible, but
Conda always names the library directory "lib". So this seems to call for a
Conda-specific patch.

This problem encountered trying to build the X11 IRAF programs, since they
still rely on xmkmf.