Update and tidy conda-py2-builder Docker setup
The main feature is that we now have a better exclude-based system for copying
recipe files to the work directory, rather than an include-based system. We also
farm out the build commands to a script in /work so that I can update the script
without needing to rebuild the container.