dockerfiles/conda-py2-builder/setup.sh: hardcoding my Anaconda channel not needed
'conda build' will automatically search your linux-64 directory for packages, and it seems better
to rely on that than hardcoding this. If I want to build something with a lot of deps on a fresh
machine, I can always manually download the packages to the linux-64 directory and run
'conda index'.