recipes/casa-nrao-casacore: get measures working on OS X
The casacore measures code involved some casa::String() objects initialized
with hardcoded paths. When building for conda, these paths are very long and
then padded with NULs when the package is installed on a person's machine. I
guess that the String() code must somehow encode the constant's length in the
compiled executable, because `String("rewritten constant") + "/foo"` would
yield only "rewritten constant", not "rewritten constant/foo" -- I infer that
the append happened at the original end of the string, after all the NULs.

To fix this I have tried to replace all forms of `String(<PREFIX-CONSTANT>)`
with `String(String(<PREFIX-CONSTANT>).chars())`, which then results in an
object at runtime that does not have all of the extra NULs. The changes in
this commit are enough to get my simple testcase to pass:

   import casac
   casac.casac.measures ().obslist ()