recipes/gtk3: dev-usability improvements
Require the version of pango that I provide, since stock Anaconda now provides
a newer version that doesn't provide the GObject Introspection files needed
for using Pango (and hence GTK+3) in Python.

Also disable the annoying message when GTK+3 fails to load a module. The issue
here, I believe, is that the loaded modules are communicated via X Windows
Atoms or something, so when you run a Conda GTK+3 app it's told that certain
modules ought to be loadable. But of course we don't include every OS vendors'
random GTK+3 modules, so we can't find them, resulting in the annoying
message. The issue ought to be harmless, though, and I get annoyed when my
programs print out warnings like that, so I'm just eliminating it.