recipes/modern-xorg-stack: update to 2016.03
Add libXcursor and util-macros to hush up compilation complaints. This may or
may not be needed to get cursor operations working in the Gtk+3 stack.

Even though I committed the changes for the gtk3 rebuild before this one, I
built an updated version of this package first, so the rebuilt packages link
against libXcursor as desired.