recipes/texlive-selected: fix pdflatex and probably notebook PDFification
pdflatex didn't work because the .fmt files embed absolute paths in a way
that's not compatible with Conda's patching -- they're binary files, but
null-padding doesn't work. So we have to rerun fmtutil in a post-link script,
which took a bit of work to get working nicely.

We also include a bunch more packages that are needed for the test case I've
been given.