recipes/texlive-selected: update for TeXLive 2016 and adding 'psnfss'
I just wanted to add 'psnfss' to get 'times.sty', but it turns out that TeXLive 2016
came out and the way that they distribute files, we have to move to the new release
if we want to build a new package.

This turned out to be a bit of a pain. Files that we used before got silently
moved into the new (?) "tex-ini-files" package. Then I ran into new problems
with the "language.dat" file (and friends), which now (?) lists every possible
hyphenation scheme in the distributed version, which led to errors on missing
files. In principle one can run "tlmgr generate" to update this file and its
friends, but that relies on the TeXLive package database which we don't
maintain. So, because it was easier, I've just started including all of the
hyphenation packages in the bundle by default. This is a bit lame since we
probably don't reeeeally need Mongolian hyphenation rules in every package,
but such is life.