Update CASA packages for newly-discovered "ncursesw" issue.
CASA explicitly links with ncurses for Readline, because apparently readline
doesn't automatically pull in the right shared library dep. But Anaconda's
ncurses package provides "libncursesw.so" for some hopefully-good reason, so
our packages ended up depending on a binary that wasn't provided in the Conda
distribution, resulting in busted shared library links. As best I can tell the
best solution is to just patch the CASA build scripts to look for this
ncursesw library rather than the w-less form.