recipes/texlive-selected: new release
This new version seems to actually work again, which is nice. It now has the
tools to generate some compiled font files on-the-fly. I'm not thrilled about
that mode but the impression that I get is that it's just a necessary part of
how things work. Once again I'm not sure why I didn't need to use this stuff
before but I can't find any evidence that I'm missing any bits that would make
it unneccessary.