Revert ecb7378: rename qt4 back to qt4dbus
I've made things kind of messy with simultaneous efforts to change a lot of
things. This branch of the repo is reserved for recipes being built with *my*
build system for my *private* packages on top of the conda-forge system.
Different branches will be used for the prototypes as these recipes are
gradually shifted over to be built within the conda-forge system, which
includes both adaptations for their builders, and higher standards of
packaging quality.

Not sure if this recipe builds as-is, but there's an existing `qt4dbus`
package in the `pkgw-forge` channel that should work, unless some dependency
got upgraded under us.