recipes/qt4: restructure to parallel the conda-forge version
This version tries to hew closely to the conda-forge version of the recipe,
with the change that we make it parallel-installable with Qt 5. Unlike
previous iterations, here we put everything inside $PREFIX/qt4, rather than
e.g. $PREFIX/lib/qt4/. This seems to be necessary for qmake and things to work
in the way that most projects expect.

This recipe should be close to submittable to conda-forge, but Qt takes way
too long to build on the CI systems, so it can only be built manually, which
is a whole huge thing I don't want to undertake right now. This version also
uses the X.org X11 libraries and we haven't quite worked through the policy
there in conda-forge, yet.