dockerfiles/forge-builder/Dockerfile: duh, need to keep this at CentOS 6
I forgot that the whole point of building on CentOS 6 is to maintain glibc
compatiblity with old OS images used on some HPC clusters. Comment added to
that effect.

I also need a newer devtoolset, though, to build CASA. It turns out that
current CentOS 6 provides devtoolset version 7 (!), which is based on GCC 7.
Probably not unrelated that the "new" conda compilers are GCC-7-based. A test
install at NRAO indicates that the resulting packages work OK.