aoflagger: update for Linux
Not updating the macOS build since the use of the C11 function
`aligned_alloc` is causing clang++ build problems. Can't update to
`-std=c++17` which would apparently fix this problem, since that
introduces its own problems. Not going to spend time on this since I'm
not aware of anyone that uses my packages that uses aoflagger on Macs.